% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataToOracle.R
\name{DataFrameToORC}
\alias{DataFrameToORC}
\title{DataFrameToORC}
\usage{
DataFrameToORC(JDBCClassPath, Url, Id, Pw, Owner, Table,
  Truncate = FALSE, DataOnR, NCore = 1, NumOnePack = 500)
}
\arguments{
\item{JDBCClassPath}{must be inputted. Where is the path to the JAR file containing the driver?}

\item{Url}{Url must be inppted. What is the DB host, port, sid?}

\item{Id}{Id must be inppted. What is the DB connecting ID?}

\item{Pw}{Pw must be inppted. What is the DB connecting password?}

\item{Owner}{Owner must be inppted. Who created the table?}

\item{Table}{Table must be inppted. What is the table name?}

\item{Truncate}{default is FALSE. TRUE is that rows from uploading table delete.}

\item{DataOnR}{DataOnR must be inppted. What is the data name which uploading in the table?}

\item{NCore}{defalut is 1. How many use local computer cores? if your all core use, it use detectCores().}

\item{NumOnePack}{defalut is 500. upload rows number once. if you select NCore 8 and NumOnePack 1000, 8000 rows insert DB once.}
}
\description{
This function allows you to upload in oracle DB
}
\examples{
DataFrameToORC(
 JDBCClassPath = "C:/ojdbc6.jar"
 , Url = "jdbc:oracle:thin:@<db_ip>:<db_port>:<db_sid>"
 , Id = "LeeYuHyung"
 , Pw = "GoodMan"
 , Owner = "LYH"
 , Table = "TestTable"
 , Truncate = FALSE
 , DataOnR = "MyData"
 , NCore = 1
 , NumOnePack = 500
)
}
\keyword{dldbgud}
